Add NewtReadPkgWithOffset(), which works like NewtReadPkg() except that it takes an offset to apply to any data lookups.
Rationale: The packages located in the REx Blocks don't contain
package-relative addresses, but absolute addresses in the ROM.
If we try to read them as-is, we'll crash as they'll exist far
outside the length of the package data.

Instead we can subtract this offset from all the p_objs, and
happily read in the package.